fix: move api errors to separate module to avoid indirect imports of auth code (#1154)
