+++ app/services/atlases.ts
@@ -1,6 +1,6 @@
 import pg from "pg";
 import { ValidationError } from "yup";
-import { InvalidOperationError } from "../../app/utils/api-handler";
+import { InvalidOperationError } from "../../app/utils/api-errors";
 import { getCrossrefPublicationInfo } from "../../app/utils/crossref/crossref";
 import {
   ATLAS_STATUS,
