+++ app/services/concepts.ts
@@ -15,7 +15,7 @@ import {
   ConflictError,
   InvalidOperationError,
   NotFoundError,
-} from "../utils/api-handler";
+} from "../utils/api-errors";
 import { mapDatabaseError } from "./database";
 
 /**
