+++ app/utils/api-handler.ts
@@ -13,6 +13,12 @@ import {
   getAtlasIdByUrlParameter,
 } from "../services/atlases";
 import { query } from "../services/database";
+import {
+  ApiError,
+  ForbiddenError,
+  NotFoundError,
+  UnauthenticatedError,
+} from "./api-errors";
 import { S3KeyFormatError } from "./files";
 
 interface UserProfile {
@@ -31,47 +37,6 @@ export type Handler = (
   req: NextApiRequest,
   res: NextApiResponse,
 ) => Promise<void>;
-
-abstract class ApiError extends Error {
-  abstract statusCode: number;
-  fieldPath: string | null;
-  constructor(message?: string, fieldPath: string | null = null) {
-    super(message);
-    this.fieldPath = fieldPath;
-  }
-}
-
-export class InvalidOperationError extends ApiError {
-  name = "InvalidOperationError";
-  statusCode = 400;
-}
-
-// Represents a request conflict (e.g., ETag mismatch/version conflicts)
-// Maps to HTTP 409 Conflict
-export class ConflictError extends ApiError {
-  name = "ConflictError";
-  statusCode = 409;
-}
-
-export class UnauthenticatedError extends ApiError {
-  name = "UnauthenticatedError";
-  statusCode = 401;
-}
-
-export class ForbiddenError extends ApiError {
-  name = "ForbiddenError";
-  statusCode = 403;
-}
-
-export class AccessError extends ApiError {
-  name = "AccessError";
-  statusCode = 400;
-}
-
-export class NotFoundError extends ApiError {
-  name = "NotFoundError";
-  statusCode = 404;
-}
 
 /**
  * Creates an API handler function that calls the provided middleware functions in order for as long as each one calls `next` function passed to it.
