+++ pages/api/me.ts
@@ -4,12 +4,12 @@ import {
 } from "../../app/apis/catalog/hca-atlas-tracker/common/entities";
 import { METHOD } from "../../app/common/entities";
 import { createUser, updateLastLogin } from "../../app/services/users";
+import { UnauthenticatedError } from "../../app/utils/api-errors";
 import {
   getActiveUser,
   getProvidedUserProfile,
   handler,
   method,
-  UnauthenticatedError,
 } from "../../app/utils/api-handler";
 
 export default handler(method(METHOD.PUT), async (req, res) => {
